split syncsetting to option page
