+++ public/popup.html
@@ -1,12 +1,15 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Document</title>
-  </head>
-  <body style="width: 450px; height: 400px; margin: 0;">
-    <div id="app"></div>
-    <script src="popup.js"></script>
-  </body>
+
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>Document</title>
+</head>
+
+<body style="width: 450px; height: 400px; margin: 0;">
+  <div id="app"></div>
+  <script src="popup.js"></script>
+</body>
+
 </html>
