+++ src/popup/main.tsx
@@ -3,17 +3,16 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import styled from 'styled-components'
 import mstyled from '@material-ui/core/styles/styled'
-import RepositoryItem from './repositoryItem'
+//import RepositoryItem from '../options/repositoryItem'
 import {
-  repositories,
   getActiveRepository,
-  setActiveRepository,
 } from '../repositories/repositoryManager'
 import { Container, Typography, Paper } from '@material-ui/core'
 import useSWR, { mutate } from 'swr'
 import { MainTheme } from '../uiTheme'
 import SyncSection from './syncSection'
 import SettingSection from './settingSection'
+import WhiteButton from '../ui/whiteButton'
 
 const Root = mstyled(Container)({
   paddingTop: '1rem',
@@ -51,11 +50,6 @@ const SubName = styled.h2`
   font-weight: normal;
 `
 
-const Repositories = mstyled(Paper)({
-  padding: '1rem',
-  marginBottom: '1rem',
-})
-
 const App = () => {
   const { data: activeRepoID } = useSWR('activeRepo', async () => {
     const repo = await getActiveRepository()
@@ -78,24 +72,6 @@ const App = () => {
             </SubName>
           </Brand>
         </Header>
-        <Repositories elevation={0}>
-          <Typography variant="h2" gutterBottom>
-            同期サービス選択
-          </Typography>
-          {repositories.map((r) => (
-            <RepositoryItem
-              key={r.id}
-              repo={r}
-              isActive={activeRepoID === r.id}
-              onRadioClick={() =>
-                mutate('activeRepo', async () => {
-                  await setActiveRepository(r.id)
-                  return r.id
-                })
-              }
-            />
-          ))}
-        </Repositories>
         <SyncSection hasActiveRepo={!!activeRepoID} />
         <SettingSection />
       </Root>
